Make bflabel/aclabelfont work across acronym versions
In the latest version of acronym (1.41), the command bflabel was
removed and replaced with aclabelfont. In classicthesis, we want
to change whichever of these is present so acronyms are printed
in normal, not bold, face.

In this change, we check whether bflabel is defined and if it
we assumes acronym <1.41 and fix the bflabel command.
If it isn't, we change aclabelfont which will work in
acronym>=1.41.